fixed warning by using static_cast

+++ src/bigquery_arrow_reader.cpp
@@ -71,7 +71,7 @@ shared_ptr<google::cloud::bigquery::storage::v1::ReadStream> BigqueryArrowReader
         throw BinderException("Read session is not initialized.");
     }
     auto next_stream_idx = next_stream++;
-    if (next_stream_idx >= read_session->streams_size()) {
+    if (static_cast<int>(next_stream_idx) >= read_session->streams_size()) {
         return nullptr;
     }
     auto stream = make_shared_ptr<google::cloud::bigquery::storage::v1::ReadStream>( //
